FX-3134 Moved helper to separate file and fixed types
